Refactor statement.go
